requirements.txt
Flask==2.3.2
Flask_SocketIO==5.3.4
python-socketio==5.8.0
shioaji==1.1.6
